Make clang-tidy work again

See merge request acubesat/obc/ecss-services!126

+++ inc/Services/FunctionManagementService.hpp
@@ -89,7 +89,7 @@ public:
 	 */
 	void include(String<ECSSFunctionNameLength> funcName, void (*ptr)(String<ECSSFunctionMaxArgLength>));
 
-	int getMapSize() {
+	size_t getMapSize() {
 		return funcPtrIndex.size();
 	}
